Add cleaning ability to the test release runner

+++ scripts/99x-build-test-release.sh
@@ -6,11 +6,15 @@ export VERSION=v2.1
 export DEPLOY_HOME=/home/bherr/workspaces/hubmap/hra-kg/dist-${VERSION}
 export EXTRA_DOs="graph/ccf/v2.3.0 graph/ds-graphs-enrichments/v2023"
 export COLLECTIONS="collection/hra/$VERSION collection/hra-api/$VERSION collection/ds-graphs/v2023"
+export CLEAN="true"
 
-rm -rf $DEPLOY_HOME
-mkdir -p $DEPLOY_HOME
-echo "*" > $DEPLOY_HOME/.gitignore
+if [ "$CLEAN" = "true" ]; then
+  rm -rf $DEPLOY_HOME
+  mkdir -p $DEPLOY_HOME
+  echo "*" > $DEPLOY_HOME/.gitignore
+fi
 
+# Determine Digital Objects to process
 touch $DEPLOY_HOME/.digital-objects
 for DO in $EXTRA_DOs; do
   echo $DO >> $DEPLOY_HOME/.digital-objects
@@ -21,30 +25,40 @@ done
 export DOs=$(sort $DEPLOY_HOME/.digital-objects | uniq)
 rm -f $DEPLOY_HOME/.digital-objects
 
+# Normalize digital objects
 for d in $DOs $EXTRA_DOs; do
-  echo
-  echo "---- START NORMALIZE ${d} ----"
-  do-processor --exclude-bad-values normalize $d
-  echo "---- END NORMALIZE ${d} ----"
-  echo
+  if [ ! -e digital-objects/$d/normalized/normalized.yaml ] || [ "$CLEAN" = "true" ]; then
+    echo
+    echo "---- START NORMALIZE ${d} ----"
+    do-processor --exclude-bad-values normalize $d
+    echo "---- END NORMALIZE ${d} ----"
+    echo
+  fi
 done
 
+# Enrich digital objects
 for d in $DOs $EXTRA_DOs; do
-  echo
-  echo "---- START ENRICH ${d} ----"
-  do-processor enrich $d
-  echo "---- END ENRICH ${d} ----"
-  echo
+  if [ ! -e digital-objects/$d/enriched/enriched.ttl ] || [ "$CLEAN" = "true" ]; then
+    echo
+    echo "---- START ENRICH ${d} ----"
+    do-processor enrich $d
+    echo "---- END ENRICH ${d} ----"
+    echo
+  fi
 done
 
+# Deploy digital objects
 for d in $DOs $EXTRA_DOs; do
-  echo
-  echo "---- START DEPLOY ${d} ----"
-  do-processor deploy $d
-  echo "---- END DEPLOY ${d} ----"
-  echo
+  if [ ! -e $DEPLOY_HOME/$d/graph.ttl ] || [ "$CLEAN" = "true" ]; then
+    echo
+    echo "---- START DEPLOY ${d} ----"
+    do-processor deploy $d
+    echo "---- END DEPLOY ${d} ----"
+    echo
+  fi
 done
 
+# Process collections
 for d in $COLLECTIONS; do
   echo
   echo "---- START COLLECTION ${d} ----"
@@ -55,6 +69,7 @@ for d in $COLLECTIONS; do
   echo
 done
 
+# Finalize deployment, including creating a fresh blazegraph.jnl
 echo
 echo "---- START FINALIZING ----"
 do-processor finalize
@@ -63,6 +78,7 @@ blazegraph-runner select --journal=$DEPLOY_HOME/blazegraph.jnl ./src/high-level-
 echo "---- END FINALIZING ----"
 echo
 
+# Sync up to S3
 echo
 echo "---- START Syncing built files to S3 ----"
 for d in $DOs $EXTRA_DOs $COLLECTIONS; do
